+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 *Features*
 
+* Adding `xan matrix count` & `xan matrix adj`.
 * Adding `front_coding` window function.
 * Timestamp support with `xan plot -LT`.
 * Adding `xan rename -n/--no-headers` support for `-p/--prefix` & `-x/--suffix`.